[Refactor] model port, adaptor 패키지 구조 변경 및 rename (#294)
* refactor: port, adapter usecase 단위로 패키지 구조 수정 및 rename

* test: 패키지 구조 수정 및 rename에 따른 TC 수정

* chore: 불필요한 어노테이션 제거
