+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/model/usecase/get_models/GetModelsUseCase.java
@@ -1,10 +1,9 @@
-package softeer.be_my_car_master.api.model.usecase;
+package softeer.be_my_car_master.api.model.usecase.get_models;
 
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.model.dto.response.GetModelsResponse;
-import softeer.be_my_car_master.api.model.usecase.port.ModelPort;
 import softeer.be_my_car_master.domain.model.Model;
 import softeer.be_my_car_master.global.annotation.UseCase;
 
@@ -12,10 +11,10 @@ import softeer.be_my_car_master.global.annotation.UseCase;
 @RequiredArgsConstructor
 public class GetModelsUseCase {
 
-	private final ModelPort modelPort;
+	private final GetModelsPort getModelsPort;
 
 	public GetModelsResponse execute() {
-		List<Model> models = modelPort.findModels();
+		List<Model> models = getModelsPort.findModels();
 		return GetModelsResponse.from(models);
 	}
 }
